Recaptcha_slasher_v1.1.2
Deployment of CNN trained to classify Google Recaptcha images.
The pipeline set not to just solve Recaptcha but to collect
images further. The accuracy of the model could be
improved (I belive, ~0.7 on test set) nevertheless the specifics
of Recaptcha Service were taken into account even when the 
model predicts insufficiant number of pictures.

During testing and deployment became apparent the necessity
to extend model to classify more categories, while training
images are collected temporary solution should be
random clicking (function random_solver)

As the algorythm does not perform good enough another
idea is to overwrite the model not just to recognize but
to detect and recognize objects as big reason of not
all that successful performance is categorical
classification despite pictures having several objects.